backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.14.1
python-dotenv==1.0.0
flask-bcrypt==1.0.1
pyjwt==2.8.0
flask-mail==0.9.1
cloudinary==1.36.0
waitress==3.0.2
torch==2.8.0
torchvision==0.23.0
pillow==11.3.0
